+++ Utilities/DebugConsole/changelog.md
@@ -1,5 +1,9 @@
 Changelog of Debug Console, a Phoenix Point Mod by Sheepy
 
+# Version ?
+
+* Add Duration declaration
+
 # Version 1.0, 2020-03-18
 
 * Enable console, toggled by `.